Support posting on multiple accounts
- This requires post_id to be a JSONField attribute
- This fixes post_id may not being accurate when the service has multiple accounts enabled
- Another change with this patch is post is only considered posted and postschedule obj is only deleted when send_post returns post_id

+++ base/scheduler.py
@@ -97,12 +97,16 @@ def post_scheduler(pending_objects, updating_objects, **kwargs):
         for account in account_objects:
             access_token = getattr(account, "access_token")
             api_base_url = getattr(account, "api_base_url")
+            uid = getattr(account, "uid")
+            # get post_id specific to account if it is an existing and format conforming post
+            account_pid = [pid.split("_")[-1] for pid in post_object.subject.post_id if pid.split("_")[0] == uid] if post_object.subject.post_id and isinstance(post_object.subject.post_id, list) else []
+            account_pid = account_pid[0] if account_pid else None
             try:
                 post_id = send_post(
                     content,
                     access_token=access_token,
                     api_base_url=api_base_url,
-                    post_id=post_object.subject.post_id,
+                    post_id=account_pid,
                     receiver=post_object.receiver,
                     visibility=post_object.visibility,
                 )
@@ -110,11 +114,17 @@ def post_scheduler(pending_objects, updating_objects, **kwargs):
                 log_error = message("LOG_EXCEPT", exception=e, verbose='Post "%s" (%s) has failed to be sent' % (post_object, post_object.name), object=post_object)
                 logger.error(log_error)
             else:
-                # update subject object post_id
-                post_object.subject.post_id = post_id
-                post_object.subject.save(update_fields=["post_id"])
+                if not post_id:
+                    return
+                pid = "%s_%s" % (uid, post_id)
+                # update subject object post_id if new or non-format conforming post
+                if not account_pid:
+                    if not isinstance(post_object.subject.post_id, list):
+                        post_object.subject.post_id = list()
+                    post_object.subject.post_id.append(pid)
+                    post_object.subject.save(update_fields=["post_id"])
                 log_message = message("LOG_EVENT", event='Post "%s" (%s) has been sent' % (post_object, post_object.name))
                 logger.info(log_message)
-            # delete post schedule object if it has been sent
-            if post_object.subject.post_id:
-                post_object.delete()
+                # delete post schedule object if it has been sent on current account
+                if isinstance(post_object.subject.post_id, list) and pid in post_object.subject.post_id:
+                    post_object.delete()
